Tranform review JSON before saving it to database
The transformation involves adding a new field named "files" to the
review JSON before saving it to the database. The "files" field is an
array of strings containing the union of paths of all files in all
patchsets in the code review.

See Issue #2 for more information on the need for the transformation.