Add method to transform reviews by adding field
The fields being added at the moment are "reviewed_files" and
"committed_files" which have values of type JSON array containing names
of files that were reviewed in the code review and those that were
eventually committed, respectively.

Using this change as an opportunity, tests were written and test data
files were also added to the repository.