Create index on reviewed_files and committed_files
'reviewed_files' and 'committed_files' are two fields that are
explicitly added during data load. The 'reviewed_files' field will
contain a list of file paths aggregated from all patch sets and the
'committed_files' field will contain a clist of file paths from the last
patch set in the review, with the assumption that the last patch set is
what is committed to the source code.