Remove the use of 'labels' JSON field for bugs
When loading bugs in to the database, the assumption made earlier was
that the CVE associated with a bug can be in either the field labeled
"cve" or "labels". However, upon closer examination, the CVE appears in
the "labels" field if and only if it appears in the "cve" field making
the use of the "labels" field redundant.