Load all bugs and reviews before mapping
The assumption made with the earlier implementation was that the bug
associated with a review is opened in the same year as the review.
However, there may be cases when the review was created in one year and
the bug was opened in an earlier year. With this change, bugs and
reviews across all years are loaded before attempting to map reviews to
bugs.

Change resolves #4.