Fix a rather naive bug in index creation
The 'operations` property of the 'Migrations' class is the list that
should contain all the migration operations. However, the bug that was
introduced earlier defined the property 'operations' twice: once for the
creation of the index for the 'reviewed_files' field and once for the
creation of the index for the 'committed_files' field. Consequently,
only the last definition of the 'operations' property was applied to the
database leading to the index for the field 'reviewed_files' from not
being created. The change should resolve the bug.