Use an iterator when loading reviews from files
The get_reviews method was reading all the reviews from all the JSON
files before returning them to be processed. As a result, the memory
usage was very high and the caller was blocked until all the reviews
were read into memory. With this change the get_reviews returns an
iterable to allow iteration through the reviews.