Change paths and database settings by environment
The paths to data files and database settings were statically configured
in settings.py and dbsettings.py files, respectively. As a result, when
testing, the paths had to be overridden. With this change, the paths and
the database settings are dynamically configure depending on the value
of the environment variable ENV. The permitted values for the ENV
variable are DEV, TEST and PROD, with DEV being the default. As of this
change, the paths are the same for DEV and PROD. The name of the
database, however, varies.