Separate concerns when loading data into database
The process of loading the bug, review, and vulnerability data into the
database got really complex really quickly. In order to simplify
maintenance, and testing, the code to load the database has not been
split into loaders and taggers packages. The modules in the loaders
package load the data into the database and the modules in the taggers
package add specific tags to data in the database after it has been
loaded.