Refactor parallel processing infrastructure
The terminology used with the parallel processing source code earlier
was misleading. The typical producer-consumer pattern is not intuitive
in the context of the usage of the parallel processing infrastructure so
the doer-aggregator pattern is used instead. The term 'doer' has the
connotation of doing work, which could be processing data or saving
records to the database. The term 'aggregator' summarizes the role
pretty effectively.