Add materialized views to cache document frequency
The current database schema does not enable the easy computation of
document frequency i.e. the number of documents (code reviews in the
context of the application) that contain a specific token or its lemma.
The creation of materalized views simplifies the computation of document
frequency leading to improved performance.

See Issue #10 for more information on the limitation that this change
aims to overcome.