Add code to avoid makemigrations generating it
For some odd reason, makemigrations generates CreateModel instances into
migrations even for models that have been specifically configured to be
unmanaged. However, the migration code created does not cause any change
to the database. Adding this boilerplate to this file as the models are
associated with the materialized views created in this file.