Fix assumption in mapping vulnerability to bugs
The initial design of the database reflected an incorrect assumption
that the mapping between vulnerability and bug is one-to-one. However,
in a separate exercise, a consequence of this wrong assumption revealed
itself as an integrity constraint violation. Upon closer inspection, it
was found that a single vulnerability can be mapped to multiple bugs.
The change fixes the assumption.

Change resolves #12.