Add dummy field named id to appease Django's ORM
Django, by default, assumes all models to have a primary key field named
'id'. In case of an unmanaged model mapped to a database view, the
assumption does not work quite well. To overcome this limitation, a
dummy field named id is added to the view during its creation. The value
of the id field is sourced from a sequence the creation code for which
is also part of this change.

Reference

http://schinckel.net/2014/09/01/postgres-view-meet-django-model/