Use multiple processes when loading the database
The change uses multiple processes when loading bugs and reviews into
the database. Furthermore, with this change a new design for using
multiple processes is implemented in which the number of items that need
to be iterated upon in parallel need not be known in advance enabling
the iteration to be over a generator object (i.e. an object returned by
a function when it executes yield).

Change is a contribution toward issue #3.