Enhance token remover to add additional filters
With this change the token remover is enhanced with the additional of
new filters that enable removal of tokens (a) with special characters in
them, (b) longer than specified length, and (c) that only have numbers
(irrespective of the kind i.e. integers, floats). Additionally, the
filter that removes tokens based on word length can be configured to
remove tokens with specific lengths.