Skip potentially flaky test
The change skips a test that is flaky as it depends on an external
resource---Rietveld. As of this change, the number of reviews for the
year 2008 is 6244 according to Rietveld but we expect it to be 6245. The
test must be refactored to account for such instances without failing,
perhaps a warning.