Adapt to modified uncertainty classifier
The uncertainty classifier was modified to avoid loading the classifier
from disk each time a prediction was to be made. As a result, an
object-oriented approach was chosen in the implementation of the
uncertainty classifier. The change adapts the client of the classifier
to work with the updated design.

In addition, the requirements.txt file was modified to now refer to the
uncertainty classifier published on PyPI instead of the one from GitHub.