Add infrastucture to analyze inline comments
The change implements a significant remodel of the database schema to
accommodate inline comments made to source code files in a code review.

As a direct consequence of the change, a large number of files had to be
modified to adapt them to work with the new schema. Also, any PEP8 violations
in the impacted files were addressed.