Refactor JSON loading to support loading from file
The load_json function read the contents of a JSON file as a string and
then attempted to load the JSON from string. The function was
implemented this way to overcome the issue with JSONs downloaded from
Rietveld having a Unicode null character which had to be replaced in the
text before attempting to load the JSON object. However, attemping to
use load_json to load a regular JSON file (one without Unicode nulls)
resulted in a JSONDecodeError. As a result, the sanitize argument to the
load_json function is now used to control the way in which a JSON is
loaded. If sanitize is True, file is read to string and then the JSON is
loaded from the string. If sanitize is False, the JSON is directly
loaded from file.