Load bugs JSONs into the database
After a lot of waiting, the bugs from Monorail have been downloaded in
their JSON formats using the Monorail API. The change switches loading
the bugs from CSV files downloaded from the Monorail's web interface
with the JSON files. The test CSV files are being removed with the
addition of the JSON files.

In addition to the change described, few refactoring changes were
implemented chief among which is the addition of tests to test
additional functions in app/lib/files.py.