Transform bug JSON to include participation
The change transforms the raw bug JSON from the Monorail API to add two
new keys---participants and contributors. The value corresponding the
participants key is a list of email address of developers who have
participated, by being on cc, in triaging the bug. The value
corresponding the contributors key is a list of email address of
developers who have contributed, by posting comments, to the process of
triaging the bug. These two lists will simplify the lookup of developers
who may be familiar with the bug and see if the bug familiarity impacts
the feedback they provide in corresponding code reviews.