Fix bug introduced in e9c70f9b
The assumption made when loading bugs in their JSON format was to not
sanitize unicode null characters in the raw JSON text before attempting
to load and save the JSON to the database. The assumption, however, was
shown to be incorrect when a load was attempted. The change fixes the
bug that was caused by the incorrect assumption.