Fix bug with pattern used to match unicode null
The pattern used to match unicode null characters would cause a problem
when there were preceeding slashes (\). The use of the pattern is to
replace unicode null character sequences (\u0000) with empty string
because of the limitation that PostgreSQL's JSON field has with handling
the unicode null character sequence. However, if there were slashes
preceding the sequence, the characters after the sequence would be
escaped after the seqyence is replaced. For example, consider the
substring "\\\u000020M". If "\u0000" is replaced, the resulting string
is "\\20M" which is interpreted as an incorrect escape sequence when
attempting to parse as JSON. The change now replaces the unicode null
character sequence and any preceeding slashes with the empty string
resulting in the string "20M" for the example shown above.