Resort to relying on sequence in unquoted response
The assumption made with the implementation of the get_parent function
was that responses to comments have the text of the comment being
responded to quoted in the response. However, several counter examples
were found for this assumption. As a result, a decision was made to
resort to relying on the sequence of comments in determining the parent
of a comment which has no quoted comment indicating the parent.

Note: The seemingly less accurate approach will be used *only*
identifying the parent using quoted text is not possible.