Create a btree index on Review.created field
In computing the review experience of developers, the created field will
be used to search for reviews that was created before the date of a
review comment being posted. The change creates a btree index on the
created field which will be used to cluster the review table helping
speed-up the query using the created field.