Upgrade Django from 1.10.1 to 1.11.3
Django 1.10.x did not allow the name of a column index to be explicitly
set. Django 1.11.x added the ability to specify the name of an index
using the Meta.indexes property. The need for naming indexes came from
the need to cluster review using the index on created column.