Add functions to compute experience using Pandas
An earlier attempt to compute experience metrics for author of a comment
using SQL led to scalability issues. The functions being added in this
change essentially mimic the behavior of SQL using Pandas data frame.
The change allows all relevant tables to be quried from the database
once and shared across multiple processes to be independently filtered
as needed. The change enables scaling horizontally, limited only by the
number of processors available.