Resolve issue where certain comment wrongly tagged
When developing the algorithm to tag comments as useful an assumption
was made that the parent of a change-indicating comment is useful. The
assumption was shown to be wrong in cases where a developer said
something in response to a review comment and then said "Done." (See
Line 195 of patch 30004 of patchset 20001 of review 893753002 for an
example). The change introduces logic to handle such cases. Essentially,
if the parent of a change-indicating comment is not posted by a
reviewer, we traverse the thread until one is found or the end of thread
is reached.