Create directory to contain the split data sets
The cliff.delta function from effsize package was found to be very slow
for computing effect sizes. As a result, a decision to use Python to
compute effect sizes was made. The directory being created will contain
the various data sets that will be used in the effect size computations.