Tag comments as useful by traversing the thread
With this change, comments will be tagged as useful if they were posted
by a reviewer and eventually lead to a "Done." by the developer. In
other words, comments are traversed, through parent relationships, and
every comment made by a reviewer is tagged as being useful with the
rationale that the useful comment could not have been possible without
the discussion in the thread.